Implement pulling repositiories
Added 'pull_repo' and 'update_repo' methods. Clones if no
repo exists otherwise pulls. Safer then constintly recloning

+++ app/controllers/git_controller.rb
@@ -3,21 +3,29 @@ require 'git'
 require 'fileutils'
 
 class GitController < ApplicationController
-
   def pull
-    conf = get_config
-    puts conf
-    if conf
-      clone_repo(conf['repo'], conf['branch'])
-    end
-    head 200, content_type: "text/html"
+    update_repo('blog.conf')
+    head 200, content_type: 'text/html'
   end
 
-  def clone_repo (repo, branch)
-    if Dir.exists?('blog content')
-      FileUtils.rm_rf('blog content')
-    end
+  def clone_repo(repo, branch)
+    FileUtils.rm_rf('blog content') if File.exists?('blog content')
     Git.clone(repo, branch, path: 'blog content/')
   end
 
+  def pull_repo(repo, branch)
+    g = Git.open("blog content/#{branch}", log: Logger.new(STDOUT))
+    g.pull(repo, branch)
+  end
+
+  def update_repo(config)
+    conf = get_config(config)
+    if conf
+      if Dir.directory?("blog content/#{conf['branch']}")
+        pull_repo(conf['repo'], conf['branch'])
+      else
+        clone_repo(conf['repo'], conf['branch'])
+      end
+    end
+  end
 end
